Main GUI done but needs work.
The main GUI is finished, but needs work. The server has not been added 
yet, but it will be before the next commit. Sometime before 9:30 AM 
June 7, 2016. Instructions on how to run the application will be added 
in the README.md file before then too. The server will be located in 
the morsecode-transmitter-server repository.